Terminal_result/requirements.txt
PyMuPDF==1.23.9
Pillow==10.2.0
torch>=2.0.0
transformers==4.40.1
timm>=0.6.13
git+https://github.com/salesforce/BLIP.git
